clean up rxjava logic
